Backend/requirement.txt
Flask
psycopg2-binary
bcrypt
sendgrid
twilio
APScheduler
Flask-SSLify
gunicorn
